split out the v3 data set types
This also moves another data interactive data set type that was missed before.
