+++ v3/src/components/graph/v2-graph-importer.ts
@@ -1,7 +1,8 @@
 import {ITileModelSnapshotIn} from "../../models/tiles/tile-model"
 import {toV3Id} from "../../utilities/codap-utils"
 import {V2TileImportArgs} from "../../v2/codap-v2-tile-importers"
-import {IGuidLink, isV2GraphComponent, v3TypeFromV2TypeIndex} from "../../v2/codap-v2-types"
+import {IGuidLink, isV2GraphComponent} from "../../v2/codap-v2-types"
+import {v3TypeFromV2TypeIndex} from "../../v2/codap-v2-data-set-types"
 import {GraphAttrRole, PrimaryAttrRole, axisPlaceToAttrRole} from "../data-display/data-display-types"
 import {kGraphIdPrefix, kGraphTileType} from "./graph-defs"
 import {PlotType} from "./graphing-types"
