+++ v3/src/data-interactive/handlers/attribute-list-handler.test.ts
@@ -1,4 +1,4 @@
-import { ICodapV2Attribute } from "../../v2/codap-v2-types"
+import { ICodapV2Attribute } from "../../v2/codap-v2-data-set-types"
 import { toV2Id } from "../../utilities/codap-utils"
 import { diAttributeListHandler } from "./attribute-list-handler"
 import { setupTestDataset } from "./handler-test-utils"
