+++ v3/src/v2/export-v2-document.test.ts
@@ -7,7 +7,7 @@ import { getSharedDataSets } from "../models/shared/shared-data-utils"
 import { getSharedModelManager } from "../models/tiles/tile-environment"
 import { getGlobalValueManager } from "../models/global/global-value-manager"
 import { toV2Id } from "../utilities/codap-utils"
-import { ICodapV2DataContext } from "./codap-v2-types"
+import { ICodapV2DataContext } from "./codap-v2-data-set-types"
 import { exportV2Document } from "./export-v2-document"
 
 import "../components/calculator/calculator-registration"
