+++ v3/src/v2/export-v2-document.ts
@@ -8,7 +8,8 @@ import { getSharedModelManager } from "../models/tiles/tile-environment"
 import { getGlobalValueManager } from "../models/global/global-value-manager"
 import { toV2Id } from "../utilities/codap-utils"
 import { exportV2Component } from "./codap-v2-tile-exporters"
-import { CodapV2Component, ICodapV2DataContext, ICodapV2DocumentJson } from "./codap-v2-types"
+import { CodapV2Component, ICodapV2DocumentJson } from "./codap-v2-types"
+import { ICodapV2DataContext } from "./codap-v2-data-set-types"
 
 interface IV2DocumentExportOptions {
   filename?: string
